[INJIMOB-3559]: Organize helper classes aligned with kotlin library
