+++ Sources/InjiVcRenderer/utils/TemplateHelper.swift
@@ -16,8 +16,8 @@ class TemplateHelper {
     func extractSVG(renderMethod: [String: Any],
                                    vcJsonString: String) throws -> [String] {
 
-        try RenderMethodHelper(traceabilityId: traceabilityId).validateSvgMustacheRenderSuite(renderMethod)
-        try RenderMethodHelper(traceabilityId: traceabilityId).validateTemplateRenderMethodType(renderMethod)
+        try validateSvgMustacheRenderSuite(renderMethod)
+        try validateTemplateRenderMethodType(renderMethod)
 
         guard let templateValue = renderMethod[Constants.TEMPLATE] as? [String: Any] else {
             throw InvalidRenderMethodException(traceabilityId: traceabilityId, className: className)
@@ -35,7 +35,7 @@ class TemplateHelper {
            }
         let digestMultibase = templateValue[Constants.DIGEST_MULTIBASE] as? String
         if let digestMultibase = digestMultibase,
-           try !DigestMultibaseHelper(traceabilityId: traceabilityId).validateDigestMultibase(svgString: templateResponse.body, digestMultibase: digestMultibase) {
+           try !validateDigestMultibase(svgString: templateResponse.body, digestMultibase: digestMultibase) {
             throw MultibaseValidationException(
                 traceabilityId: traceabilityId,
                 className: String(describing: Self.self),
@@ -54,5 +54,116 @@ class TemplateHelper {
             return [templateResponse.body]
         }
     }
+    
+    func validateDigestMultibase(svgString: String, digestMultibase: String) throws -> Bool {
+        guard digestMultibase.hasPrefix("u") else {
+            throw MultibaseValidationException(traceabilityId: traceabilityId, className: className,  exceptionMessage: "digestMultibase must start with 'u'")
+        }
+        
+        let encodedPart = String(digestMultibase.dropFirst())
+        guard let decoded = base64UrlNoPadDecode(encodedPart) else {
+            throw MultibaseValidationException(traceabilityId: traceabilityId, className: className,  exceptionMessage: "Base64 Decoding error")
+        }
+        
+        guard decoded.count == 34 else {
+            throw MultibaseValidationException(traceabilityId: traceabilityId, className: className,  exceptionMessage: "Invalid multihash length")
+        }
+        
+        guard decoded[0] == 0x12 && decoded[1] == 0x20 else {
+            throw MultibaseValidationException(traceabilityId: traceabilityId, className: className,  exceptionMessage: "Unsupported multihash prefix")
+        }
+        
+        let expectedHash = decoded.subdata(in: 2..<34)
+        let actualHash = Data(SHA256.hash(data: Data(svgString.utf8)))
+        
+        return expectedHash == actualHash
+    }
+    
+    private func base64UrlNoPadDecode(_ string: String) -> Data? {
+        var base64 = string
+            .replacingOccurrences(of: "-", with: "+")
+            .replacingOccurrences(of: "_", with: "/")
+        
+        let padding = 4 - (base64.count % 4)
+        if padding < 4 {
+            base64.append(String(repeating: "=", count: padding))
+        }
+        
+        return Data(base64Encoded: base64)
+    }
+    
+    func parseRenderMethod(_ jsonObject: [String: Any]) throws -> [[String: Any]] {
+        guard let renderMethodValue = jsonObject[Constants.RENDER_METHOD] else {
+            throw InvalidRenderMethodException(traceabilityId: traceabilityId, className: className)
+        }
+        
+        if let array = renderMethodValue as? [[String: Any]] {
+            if array.isEmpty || array.contains(where: { $0.isEmpty }) {
+                throw InvalidRenderMethodException(traceabilityId: traceabilityId, className: className)
+            }
+            return array
+        } else if let dict = renderMethodValue as? [String: Any] {
+            if dict.isEmpty {
+                throw InvalidRenderMethodException(traceabilityId: traceabilityId, className: className)
+            }
+            return [dict]
+        } else {
+            throw InvalidRenderMethodException(traceabilityId: traceabilityId, className: className)
+        }
+    }
+    
+    private func isSvgMustacheRenderSuite(_ renderMethod: [String: Any]) -> Bool {
+        let renderSuite = renderMethod[Constants.RENDER_SUITE] as? String ?? ""
+        return renderSuite == Constants.SVG_MUSTACHE
+    }
+    
+    private func isTemplateRenderMethodType(_ renderMethod: [String: Any]) -> Bool {
+        let type = renderMethod[Constants.TYPE] as? String ?? ""
+        return type == Constants.TEMPLATE_RENDER_METHOD
+    }
+    
+    
+    func validateSvgMustacheRenderSuite(_ renderMethod: [String: Any]) throws {
+        if !isSvgMustacheRenderSuite(renderMethod) {
+            throw InvalidRenderSuiteException(traceabilityId: traceabilityId, className: className)
+        }
+    }
+    
+    func validateTemplateRenderMethodType(_ renderMethod: [String: Any]) throws {
+        if !isTemplateRenderMethodType(renderMethod) {
+            throw InvalidRenderMethodTypeException(traceabilityId: traceabilityId, className: className)
+        }
+    }
+    
+    func parseVcJson(vcJsonString: String) throws -> [String: Any] {
+       do {
+           guard let data = vcJsonString.data(using: .utf8) else {
+               throw VcRendererException(
+                errorCode: VcRendererErrorCodes.invalidVcJson,
+                   message: "Invalid JSON input (data encoding failed)",
+                   className: String(describing: InjiVcRenderer.self),
+                   traceabilityId: traceabilityId
+               )
+           }
+           
+           guard let parsed = try JSONSerialization.jsonObject(with: data) as? [String: Any] else {
+               throw VcRendererException(
+                   errorCode: VcRendererErrorCodes.invalidVcJson,
+                   message: "Invalid JSON input (not a dictionary)",
+                   className: String(describing: InjiVcRenderer.self),
+                   traceabilityId: traceabilityId
+               )
+           }
+           
+           return parsed
+       } catch {
+           throw VcRendererException(
+            errorCode: VcRendererErrorCodes.invalidVcJson,
+               message: "Invalid JSON input (\(error.localizedDescription))",
+               className: String(describing: InjiVcRenderer.self),
+               traceabilityId: traceabilityId
+           )
+       }
+   }
     
 }
